chore(personhog): refactors service file
